backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
easyocr==1.7.0
Pillow==10.1.0
opencv-python==4.8.1.78
python-dotenv==1.0.0